Conflicts:
	perl/openprint/Estimating/Folding.pm
	perl/openprint/Estimating/Printing.pm
	perl/openprint/PaperPrice.pm
	perl/openprint/RFIDTag.pm
	perl/openprint/employee_production.pm
	www/public_html/email_content/forgotten_password.html
	www/public_html/main/project/bind/folding.html
	www/public_html/skins/PointOne Graphics Inc/css/employee.css
	www/public_html/skins/PointOne Graphics Inc/main/company/contact.html

+++ perl/openprint/press_schedule.pm
@@ -117,7 +117,11 @@ sub add {
 sub remove {
 	my ( $p_id, $s_id ) = @_;
 	foreach my $Job ( openprint::ScheduledJob::find('project_id'=>$p_id, ( $s_id ? ('service_id'=>$s_id) : () ) ) ) {
-		$Job->delete();
+		if ( $Job->service_id() > 1 ) {
+			$Job->save({'service_id'=>[ sets::exclude( [ $s_id ], $Job->service_id() ) ]});
+		} else {
+			$Job->delete();
+		} # end if
 	} # end foreach
 } # end sub remove
